+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/). This project use
 - **E1000 support** tested against an E1000
 - **Transport selection** allow user to explicitly specify whether to use local, nolocal, rest-only or all transports for the given request (where appropriate), mostly for testing/debugging
 - **Set control via REST** added ability to set a control value via REST interface.
+- **Compute real charge/discharge times** based on the battery capacity and real-time current, compute the estimated time to charge/discharge.
 
 ### Improved
 - Use device-cached `controls` mapping for local TTLV id→property conversion rather than fixed mapping (which was wrong for E1000).